+++ fractal/components/cards/CardFooter/CardFooter.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+    "title": "Card Footer",
+    "status": "ready",
+    "collated": true,
+    "preview": "@card-preview",
+    "context": {
+        "className": "test-class-name"
+    }
+};