+++ fractal/components/cards/CardFooterLink/CardFooterLink.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+    "title": "Card Footer Link",
+    "status": "ready",
+    "collated": true,
+    "preview": "@card-preview",
+    "context": {
+        "to": "http://www.google.com",
+        "label": "Test Footer Link"
+    }
+};