+++ fractal/components/cards/CardHeader/CardHeader.config.js
@@ -1,0 +1,15 @@
+module.exports = {
+    "title": "Card Header",
+    "status": "ready",
+    "collated": true,
+    "preview": "@card-preview",
+    "context": {
+        "title": "Test Header Title",
+        "subtitle": "Test Header Subtitle"
+    },
+    "variants": [
+        {"name": "card-header-simple", "context": {
+            "simple": true
+        }}
+    ]
+};