+++ fractal/components/cards/CardMedia/CardMedia.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+    "title": "Card Media",
+    "status": "ready",
+    "collated": true,
+    "preview": "@card-preview",
+    "context": {
+        "title": "Test Media Title",
+        "subtitle": "Test Media Subtitle",
+        "imageUrl": "https://placeimg.com/640/480/animals"
+    }
+};