+++ fractal/components/cards/SimpleCard/SimpleCard.config.js
@@ -1,0 +1,13 @@
+module.exports = {
+    "title": "Simple Card",
+    "status": "ready",
+    "collated": true,
+    "preview": "@card-preview",
+    "context": {
+        "title": "Test Simple Card Title",
+        "subtitle": "Test Simple Card Subtitle",
+        "text": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque non dolor interdum, dapibus lectus nec, ullamcorper enim. Duis consequat felis metus, at suscipit est aliquet et.",
+        "linkTo": "http://www.google.com",
+        "linkLabel": "Simple Card Link"
+    }
+};