+++ CHANGELOG.md
@@ -75,3 +75,9 @@ Added drawer, site header/footer bars, cleaned styles
 Still working out responsive header, content and footer templates and styles
 
 Updated styles to use a base-size-unit variable as design has changed from 10px base spacing units to 8px.  Doing this to be flexible.
+
+-----
+
+6/21/2017
+
+- Adding React Card components